+++ src/CvInPort.h
@@ -33,7 +33,7 @@ public:
    int _position;
 
 protected:
-   double update(uint64_t time) override;
+   float update(uint64_t time) override;
 };
