+++ src/Potentiometer.h
@@ -30,7 +30,7 @@ public:
    ~Potentiometer() override;
 
 protected:
-   double update(uint64_t time) override;
+   float update(uint64_t time) override;
 
 private:
    int _position;
